requirements.txt
pywebview
requests
urllib3
sounddevice
numpy
scipy
faster-whisper